Integration test for editing POMs
Adds a new integration test for editing a POM, although it is currently
only MVP and will fail on the final save.

To make sure that multiple integration tests don't causes issues with
login state (and as they are unordered) the sign_in_and_go function can
be used multiple times as it will return early if the user is already
signed in (after navigating to the link).  Unfortunately this
dramatically slows down the SSO sign in as by default Capybara find()
calls wait for a period of time for the element to appear.  If we don't
use a lot of JS for the display of elements on the page we can reduce
the max_wait_time.

Also adds a breadcrumb helper so we can check on each page that it shows
the correct number of breadcrumbs, or includes a backlink.