Simple paging tests
These tests simply check that we are able to visit each allocation tab
and navigate with the pagination links, and then the Next/Previous
links.  This test isn't exhaustive as we have more than 1150 pages to
traverse, and no First/Last links yet.  If they are added we can
paginating at the end of the page list.