Aad time for save to complete
When creating an allocation there is a delay of a few seconds whilst the
record is saving to the database.  As such this is causing the test to
fail as it's trying to move onto the next step to quickly.  I have added
a 5 second sleep to allow the record to save so the test can move on
successfully.