Remove Caseload page checks from smoke tests
We currently have smoke tests that run against production and check that
all the pages can be visited.  However, a while ago the dashboard view
was amended so only POMs can see the 'Your Caseload' / 'New Cases'
section.  As none of us can be POMs in prod NOMIS (as we'll show up in
real prisons which would cause issues) we need to remove this part of
the check as we can't actually visit these pages.