Update integration tests so they pass
We temporarily disabled the integration tests to update the case
information section during the 'create allocation' spec due to changes
being made to the case info page.  However, this work has been done and
now the tests are failing as there aren't any cases ready for
allocation.  This tiny PR enables the updating of case info again so we
can get the tests passing again.