Update allocation spec
A change to the codebase from this PR
(https://github.com/ministryofjustice/offender-management-allocation-manager/pull/812)
means that the create_allocation_spec started failing, this PR updates
the expected page path following allocation to ensure it now passes.

In addition we are now making use of geckodriver and we need to add the
geckodriver-helper gem in order for these integration tests to run
locally.